feat: Added order management for diners and merchants

+++ models/menu_item.go
@@ -8,7 +8,7 @@ type MenuItem struct {
 	gorm.Model
 	Name         string `json:"name" gorm:"not null"`
 	Description  string `json:"description" `
-	PriceInCents uint   `json:"price_in_cents"`
+	PriceInCents int64  `json:"price_in_cents"`
 	Category     string `json:"category" gorm:"index"`
 	VenueId      uint   `json:"venue_id" gorm:"not null"`
 	Venue        Venue  `json:"-"`
